started teleporters
Basically the teleporters are all set up to work with the map (check
display where it spawns the entities to see what we need to do to the
maps) but they wont do anything yet when you land on top of them.
Methods exist to help start this, probably about 40% done.