Added warp support in the map creator
Can't place, edit, or delete them yet (same with player spawns, except
for the first one), but the placeholders are there and they
load/draw/save.

I think I'll add a popup when placing a player spawn, with an option for
the spawn number. I'll also probably add a "delete warp/player spawn"
that would work in the same way as the "delete tiles".